+++ src/components/manage/Widgets/QuerystringWidget.jsx
@@ -20,6 +20,7 @@ import { defineMessages, injectIntl } from 'react-intl';
 import { getQuerystring } from '@plone/volto/actions';
 import Select from 'react-select';
 import { Icon } from '@plone/volto/components';
+import { format, parse } from 'date-fns';
 
 import clearSVG from '@plone/volto/icons/clear.svg';
 
@@ -155,7 +156,11 @@ class QuerystringWidget extends Component {
       case 'DateWidget':
         return (
           <Form.Field width={4}>
-            <Input type="date" {...props} />
+            <Input
+              type="date"
+              {...props}
+              value={format(parse(row.v), 'YYYY-MM-DD')}
+            />
           </Form.Field>
         );
       case 'DateRangeWidget': // 2 date inputs
@@ -165,7 +170,7 @@ class QuerystringWidget extends Component {
               <Input
                 type="date"
                 {...props}
-                value={row.v[0]}
+                value={format(parse(row.v[0]), 'YYYY-MM-DD')}
                 onChange={data =>
                   this.onChangeValue(index, [data.target.value, row.v[1]])
                 }
@@ -175,7 +180,7 @@ class QuerystringWidget extends Component {
               <Input
                 type="date"
                 {...props}
-                value={row.v[1]}
+                value={format(parse(row.v[1]), 'YYYY-MM-DD')}
                 onChange={data =>
                   this.onChangeValue(index, [row.v[0], data.target.value])
                 }